// src/types/PriorityTypes.ts
export enum PriorityTypes {
    NEWS = 'NEWS',                    // Novidades
    POPULARITY = 'POPULARITY',        // Mais vendidos
    NEWEST = 'NEWEST',                // Mais recente
    OLDEST = 'OLDEST',                // Mais antigo
    HIGH_PRICE = 'HIGH_PRICE',        // Maior preço
    LOW_PRICE = 'LOW_PRICE',          // Menor preço
}